added getPrice and getLiquidity
